fix(ios): coverage proof decode + ZK wiring
- Decode CoverageProofGenerateResponse with optional proofHash/verified for older API payloads
- Pass UInt32 literals to generateCoverageProof in HealthSync debug UI
- Merkle genesis / attestation service updates (ZK e2e wiring)

Made-with: Cursor

+++ AmachHealth/Sources/Views/HealthSyncView.swift
@@ -450,9 +450,9 @@ struct HealthSyncView: View {
                         let generated = try await AmachAPIClient.shared.generateCoverageProof(
                             walletAddress: address,
                             encryptionKey: key,
-                            startDayId: 1,
-                            endDayId: 36500,
-                            minDays: 20
+                            startDayId: UInt32(1),
+                            endDayId: UInt32(36500),
+                            minDays: UInt32(20)
                         )
                         let verify = try await AmachAPIClient.shared.verifyCoverageProof(
                             proof: CoverageProof(
